store: refactoring, move actions to their own files, move code around in 2pc.go
